Tweaks to the ship
Vectors are now a magnitude + angle, rather than being a dX and dY.
This makes it easier for us to start manipulating the angles when it comes
to acceleration (magnitude++) or turning (angle++).

For now, the Ship position has been made mutable as i couldn't work out
how else to set it from the render frames.

The ship will move on its own, with the arrow keys only acting to add
additional forces (accel/decel and turn)