Fix keyboard lag
Instead of responding to keydown to add acceleration, we have a new
"Delta" parameter on the ship, representing the current acceleration.
As long as the relevant key is down, the acceleration is constant and
applied on frame renders.

A small hack required here in Program.fs, where we call the
Keyboard.GetState() to get the state set up then immediately ignore it.
This is because it takes a while to load the state the first time, but
is relatively fast afterwards. If we didn't have this, then the first
load would be on the user keypress which takes a while to kick in properly